+++ Code/Tools/Inspector/ReflectionWidget.cpp
@@ -31,7 +31,7 @@ void ezQtReflectionWidget::ResetStats()
     Headers.append(" Size (Bytes) ");
     Headers.append(" Plugin ");
 
-    TypeTree->setColumnCount(Headers.size());
+    TypeTree->setColumnCount(static_cast<int>(Headers.size()));
     TypeTree->setHeaderLabels(Headers);
     TypeTree->header()->show();
   }
