+++ Code/Tools/Libs/GuiFoundation/Widgets/Implementation/ColorGradientWidget.cpp
@@ -370,7 +370,7 @@ void ezQtColorGradientWidget::PaintCoordinateLines(QPainter& p)
     lines.push_back(QLine(QPoint(xPos, area.top()), QPoint(xPos, area.top() + iLineHeight)));
     lines.push_back(QLine(QPoint(xPos, area.bottom()), QPoint(xPos, area.bottom() - iLineHeight)));
   }
-  p.drawLines(lines.data(), lines.size());
+  p.drawLines(lines.data(), static_cast<int>(lines.size()));
   p.restore();
 }
